add London and Berlin

+++ WPFMeteo/Wind.cs
@@ -8,7 +8,7 @@ namespace WPFMeteo
 {
     class Wind
     {
-        public int speed { get; set; }
+        public double speed { get; set; }
         public int deg { get; set; }
     }
 }
